// UI/DBAapp/src/app/models/server.ts
export class Server {
  Id: number;
  Hostname: string;
  IpAdresa: string;
  OS: string;
  BrojJezgara: number;
  Lokacija: string;
  IdOkruzenja: number;
  Virtuelan: boolean;
  Aktivan: boolean;
  DatumInstalacije: string;
  Klaster: string;
  RAM_GB: number;
  Storage_GB: number;
  Status: string;
  Napomena: string;
}